requirements.txt
Flask
werkzeug
pdfminer.six
pdfplumber
spacy
nameparser